pyproject.toml
[tool.poetry]
name = "kivy-opencv-zbar-scanner"
version = "0.1.0"
description = "Kivy OpenCV ZBar Scanner"
authors = ["Alex <globophobe@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/globophobe/kivy-opencv-zbar-scanner"
repository = "https://github.com/globophobe/kivy-opencv-zbar-scanner"
keywords = ["django", "admin", "theme"]
classifiers = [
    "Framework :: Django",
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
]
packages = [{ include = "simplecamera" }]

[tool.poetry.dependencies]
python = ">=3.8"
kivy = "*"
pillow = "*"
kivymd = "*"
pyzbar = "*"
opencv-python-headless = "*"
appdirs = "^1.4.4"
jsondiff = "^1.3.0"

[tool.poetry.dev-dependencies]
ipython = "*"
pyinstaller = "*"
flake8 = "*"
black = "*"
isort = "*"
invoke = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
